Cleaned up ASQRS Detect, generated profreports
Want to keep ASQRSDetect relatively clean and organized. Had some issues
with separating getRPeak and printECGReport, see errors listed in the
description on those files. Both had issues with the input variable
types though. I extended the functionality of this code andmade it
easier to vary the buffer length. My next goal is to partition the data
out and recalculate the mean every so many data points. Ultimate goal is
to calculate moving averages and SDs.