Implemented getPPeak, getTPeak, continued work on C++
getPPeak, getTPeak are very basic still, and need to incorporate some
noise calculations to become more robust. C++ code is definitely tricky,
a lot of functions that work natively in MATLAB are difficult or
impossible to implement in C++, so I'm figuring out work arounds or
writing the functions from scratch, which is time consuming.